+++ ldaptive/pom.xml
@@ -29,6 +29,10 @@
       <groupId>com.google.guava</groupId>
       <artifactId>guava</artifactId>
     </dependency>
+    <dependency>
+      <groupId>org.jetbrains.kotlin</groupId>
+      <artifactId>kotlin-stdlib-jdk8</artifactId>
+    </dependency>
 
     <dependency>
       <groupId>org.springframework</groupId>
