+++ ldaptive/src/main/java/org/ccci/idm/user/ldaptive/LdapGroup.kt
@@ -1,6 +1,7 @@
-package org.ccci.idm.user
+package org.ccci.idm.user.ldaptive
 
 import com.google.common.base.MoreObjects
+import org.ccci.idm.user.Group
 import javax.annotation.concurrent.Immutable
 
 @Immutable
