+++ ldaptive/src/main/java/org/ccci/idm/user/ldaptive/dao/mapper/AbstractUserLdapEntryMapper.java
@@ -63,9 +63,9 @@ import com.google.common.base.Strings;
 import com.google.common.base.Throwables;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Sets;
-import org.ccci.idm.user.Dn;
 import org.ccci.idm.user.Group;
 import org.ccci.idm.user.User;
+import org.ccci.idm.user.ldaptive.Dn;
 import org.ccci.idm.user.ldaptive.dao.io.ReadableInstantValueTranscoder;
 import org.ccci.idm.user.ldaptive.dao.util.DnUtils;
 import org.jetbrains.annotations.Contract;
